Ensure atomic fields remain atomic.
Bindgen currently doesn't handle atomic fields correctly. To work around
this issue, this commit adds some postprocessing after the bindings have
been generated: atomic fields of structs and unions are detected in
julia.h, this information is used to update the generated bindings.

Marking static data with `#link[name = "libjulia"]` in the Windows
bindings is also included in this postprocessings, which gets rid of the
cryptic sed invocation in generate_bindings.sh