Only share ledger between jlrs instances
Libraries written in Rust that have been exposed with the julia_module
macro might use different versions of jlrs and might have been compiled
with different versions of Rust. This means foreign types should only be
used in the crate that defines them, and each package needs its own
stack type and thread pool.

The ledger is now implemented in its own crate and distributed as a JLL.
Data is now tracked on the pointer rather than the range-level. This
gets rid of a bunch of awkwardness related to resizing arrays.